Initial Commit
Initial code for the safety stop feature. Implemented  in motors.c
because this is a motor feature. Keeps main.c clean too. Reorganized the
file a little bit, put the statics on top, externs on the bottom
(because of include this is useful). Not tested, but timer should fire
only if haven't set the motor speed in last 5 seconds.